giac: fix `#include <giac/giac.h>` in musl
As reported in #45708, including <giac/giac.h> fails in musl because the
HAVE_TGAMMAF macro from autotools is not available at this time.

Here we include a patch adapted from
https://github.com/geogebra/giac/commit/618a5de3349f8def5bdb909fea42fb447b6aad4a
which is the solution upstream will include in next release of giac.

We need to fix this now so we can build sagemath in musl!