musl: remove compat basename declaration and add RLIMIT_RTTIME define
This declaration was added to musl to avoid warnings for applications that
expect GNU basename() in string.h.

Pre-ansi style declarations were used to avoid causing errors due to
incompatibile declarations of basename().

In C23, just like in C++, function declarations with no arguments are treated
as having no arguments, instead of having unspecified arguments. As such,
this creates a conflicting declaration in C23.